# [Performance] Cache feature path loaded result state