When loading ruby files, the `Truffle::FeatureLoader` service checks if a load path exist for loaded features. This comes with comparing a lot of strings and usually redundant. This PR cached the lookup result. 

Note: the lookup is based on the global `$LOAD_PATH` variable - which can change. We're flushing the cache each time this happens.

The results are small, though noticeable.

The number of times the actual function (`Truffle::FeatureLoader.feature_path_loaded?`) called (`--cputracer`) has been decreased to 28% (= 72% cache hit):

Before:
```
 Truffle::FeatureLoader.feature_path_loaded? | 3250   0.9% | 943  29.0% | 2307  71.0% | ../../../../truffle/truffleruby/src/main/ruby/truffleruby/core/truffle/feature_loader.rb~201-213:7030-7565
```

After
```
 Truffle::FeatureLoader.feature_path_loaded? | 925   0.3% | 642  69.4% | 283  30.6% | ../../../../truffle/truffleruby/src/main/ruby/truffleruby/core/truffle/feature_loader.rb~218-230:7770-8305
```

This makes a tiny improvement in time (12% improvement):

Before:
```
 Truffle::FeatureLoader.feature_path_loaded? || 82ms   0.7% || 82ms   0.7% || ../../../../truffle/truffleruby/src/main/ruby/truffleruby/core/truffle/feature_loader.rb~201-213:7030-7565
```

After:
```
 Truffle::FeatureLoader.cached_loaded_feature_path || 72ms   0.9% || 39ms   0.5% || ../../../../truffle/truffleruby/src/main/ruby/truffleruby/core/truffle/feature_loader.rb~203-213:7095-7612
```

To zoom out a bit, the main parent call (3% improvement):

Before:

```
 Truffle::FeatureLoader.find_feature_or_file || 3480ms  29.3% || 50ms   0.4% || ../../../../truffle/truffleruby/src/main/ruby/truffleruby/core/truffle/feature_loader.rb~79-141:2566-4768
```

After:

```
 Truffle::FeatureLoader.find_feature_or_file || 3386ms  40.8% || 33ms   0.4% || ../../../../truffle/truffleruby/src/main/ruby/truffleruby/core/truffle/feature_loader.rb~84-146:2804-5006
```

All measurements are from a mid-sized Rails project test execution: `jt --use jvm-ce ruby -S ./bin/rails test test/...`