@itarato 2023-04-25
I looked into re-using `@loaded_features_index` but the data structure didn't really look fitting for this.

Also - does this change worth pushing at all?