@eregon 2023-04-26
Could you try https://github.com/oracle/truffleruby/pull/3010#discussion_r1172493598 ?
That is, `$LOAD_PATH` as a Set/Hash, so `include?` is O(1).
Seems cleaner as a separate PR.

The approach in this PR seems suboptimal because it would result in one item per `loaded_feature` (i.e. per element in `$LOADED_FEATURES`, actually even if not there) so quite a lot of memory. And also it might need to check the `$LOADED_FEATURES` version too or get invalidated if e.g. that file gets loaded. a Set/Hash of `$LOAD_PATH` doesn't have any of these drawbacks.