@itarato 2023-05-09
Just an update - haven't forgot about this one. I have a [toy branch](https://github.com/oracle/truffleruby/compare/master...itarato:truffleruby:understand_feature_loader_agressive) where I reduced a lot of complexity and added caching to a few places. The fixes are:
- `find_file` is called redundantly on the same file (also itself being a significant part of where time is spent) - reduced it by only loading a file once
- reducing redundant logic in `find_feature_or_file` (that's a quite drastic simplification, however it follows what should happen in logic)
- removing a few other redundant ops (eg double checking encoding, null bytes, etc)

The problem is, even with lower node and call counts, the times are the same or worse. Which I presume can be explained with it being in the cold phase (first 1 min at least). I even run everything with no compilation but even then finding something that improves was hard.

I'll keep this draft (or maybe just close it) until I have something that actually helps.